메일 엔티티,dto 수정2차

+++ src/main/java/com/example/companycore/controller/mail/DynamicMailPreviewController.java
@@ -146,7 +146,7 @@ public class DynamicMailPreviewController {
 
         Optional<ButtonType> result = confirmationAlert.showAndWait();
         if (result.isPresent() && result.get() == ButtonType.OK) {
-            Long messageId = selectedMessage.getMessageId();
+            Integer messageId = selectedMessage.getMessageId();
             User user = apiClient.getCurrentUser();
             Long userId = user.getUserId();
